................
  r168613 | murf | 2009-01-14 13:51:26 -0700 (Wed, 14 Jan 2009) | 9 lines
  
  Merged revisions 168608 via svnmerge from 
  https://origsvn.digium.com/svn/asterisk/branches/1.4
  
  ........
    r168608 | murf | 2009-01-14 12:34:35 -0700 (Wed, 14 Jan 2009) | 1 line
    
    app_page was failing to compile in dev-mode on my gcc-4.2.4 system. This change gets rid of the warning.
  ........
................


git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.6.1@168621 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_page.c
@@ -116,7 +116,7 @@ static int page_exec(struct ast_channel *chan, void *data)
 	}
 
 	if (!(dial_list = ast_calloc(num_dials, sizeof(void *)))) {
-		ast_log(LOG_ERROR, "Can't allocate %ld bytes for dial list\n", (sizeof(void *) * num_dials));
+		ast_log(LOG_ERROR, "Can't allocate %ld bytes for dial list\n", (long)(sizeof(void *) * num_dials));
 		return -1;
 	}
